# Conflicts:
#	prboom2/VisualC2015/Doom.vcxproj
#	prboom2/VisualC2015/Doom.vcxproj.filters
#	prboom2/configure.ac
#	prboom2/src/Makefile.am
#	prboom2/src/info.h
#	prboom2/src/scanner.cpp
#	prboom2/src/umapinfo.cpp

+++ prboom2/src/g_overflow.c
@@ -37,6 +37,7 @@
 
 #include <stdio.h>
 #include <stdarg.h>
+#include <stdint.h>
 
 #include "g_overflow.h"
 
@@ -172,7 +173,7 @@ void InterceptsOverrun(int num_intercepts, intercept_t *intercept)
 
       InterceptsMemoryOverrun(location, intercept->frac);
       InterceptsMemoryOverrun(location + 4, intercept->isaline);
-      InterceptsMemoryOverrun(location + 8, (int) intercept->d.thing);
+      InterceptsMemoryOverrun(location + 8, (int)(intptr_t) intercept->d.thing);
     }
   }
 }
